refactor: Tweaks on Media Card Options Button
New Transition, Removed MediaActionOptionsButton component.module.css file, refactored MediaActionOptionsButton component, and made changes to various other files including component.module.css, index.tsx, and anilist.ts.
