+++ app/components/MediaCards/MediaCard/MediaActionOptionsButton/OptionsPanel/component.module.css
@@ -26,6 +26,8 @@
 
 .options_panel {
 
+    cursor: default;
+
     background-color: var(--white-100);
 
     overflow: auto;
@@ -58,6 +60,8 @@
 
         width: 320px;
         max-height: 500px;
+
+        box-shadow: 0px 3px 8px 6px #00000080;
 
         top: 40px;
         bottom: auto;
