+++ app/components/MediaCards/MediaCard/component.module.css
@@ -138,6 +138,14 @@
 /* OPTIONS BUTTON */
 .media_item_container .options_btn_container {
 
+    display: none;
+
+}
+
+.media_item_container:hover .options_btn_container {
+
+    transition: all ease-in-out 100ms;
+
     display: flex;
     flex-direction: column;
 
@@ -161,6 +169,8 @@
 
 .options_btn_container button {
 
+    transition: all ease-in-out 100ms;
+
     background-color: var(--black-100);
 
     border-radius: 50%;
@@ -173,6 +183,18 @@
     display: flex;
     justify-content: center;
     align-items: center;
+
+}
+
+.options_btn_container button:hover {
+
+    transform: scale(1.2);
+
+}
+
+.options_btn_container button[data-active=true] {
+
+    background-color: var(--white-100);
 
 }
