+++ app/lib/dataConstants/anilist.ts
@@ -1,4 +1,6 @@
-export const userMediaStatusEntries = [
+import { ApiDefaultResult } from "@/app/ts/interfaces/apiAnilistDataInterface";
+
+export const userMediaStatusEntries: { name: string, value: ApiDefaultResult["mediaListEntry"]["status"] }[] = [
     { name: "Completed", value: "COMPLETED" },
     { name: "Watching", value: "CURRENT" },
     { name: "Planning", value: "PLANNING" },
